+++ readme.txt
@@ -1,6 +1,4 @@
 refactoring
 --------------
 
-basic functionality is OK.
-but running slow & average revenue is a little bit higher (avg. sell & avg. winner are fine.)
-need figure it out
+done
